+++ packages/cht/src/Adaptor.js
@@ -130,4 +130,5 @@ export {
   lastReferenceValue,
   merge,
   sourceValue,
+  as
 } from '@openfn/language-common';