+++ packages/commcare/src/Adaptor.js
@@ -465,4 +465,5 @@ export {
   lastReferenceValue,
   merge,
   sourceValue,
+  as
 } from '@openfn/language-common';