+++ packages/mtn-momo/src/Adaptor.js
@@ -95,4 +95,5 @@ export {
   scrubEmojis,
   sourceValue,
   util,
+  as
 } from '@openfn/language-common';