+++ packages/openimis/src/Adaptor.js
@@ -119,4 +119,5 @@ export {
   lastReferenceValue,
   merge,
   sourceValue,
+  as
 } from '@openfn/language-common';