+++ packages/pesapal/src/Adaptor.js
@@ -124,4 +124,5 @@ export {
   lastReferenceValue,
   merge,
   sourceValue,
+  as
 } from '@openfn/language-common';