+++ packages/primero/src/Adaptor.js
@@ -917,4 +917,5 @@ export {
   lastReferenceValue,
   merge,
   sourceValue,
+  as
 } from '@openfn/language-common';