+++ tools/generate/template/src/Adaptor.js
@@ -103,4 +103,5 @@ export {
   scrubEmojis,
   sourceValue,
   util,
+  as
 } from '@openfn/language-common';