+++ packages/googledrive/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-googledrive
 
+## 1.0.2 - 20 June 2025
+
+### Patch Changes
+
+- Updated dependencies \[28c2e8b]
+  - @openfn/language-common@2.5.0
+
 ## 1.0.1 - 22 April 2025
 
 ### Patch Changes
@@ -8,6 +15,6 @@
 - Updated dependencies \[13bf08f]
   - @openfn/language-common@2.4.0
 
-## 1.0.0
+## 1.0.0 - 17 April 2025
 
 Initial release.