+++ packages/kobotoolbox/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-kobotoolbox
 
+## 4.1.1 - 20 June 2025
+
+### Patch Changes
+
+- Updated dependencies \[28c2e8b]
+  - @openfn/language-common@2.5.0
+
+## 4.1.0 - 06 June 2025
+
+### Minor Changes
+
+- 2123876: Add support for `sort` and `start` option in `getSubmissions`
+  function
+
 ## 4.0.0 - 30 May 2025
 
 ### Major Changes