+++ packages/mojatax/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-mojatax
 
+## 1.0.10 - 20 June 2025
+
+### Patch Changes
+
+- Updated dependencies \[28c2e8b]
+  - @openfn/language-common@2.5.0
+
 ## 1.0.9 - 22 April 2025
 
 ### Patch Changes