+++ packages/mpesa/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-mpesa
 
+## 1.0.4 - 20 June 2025
+
+### Patch Changes
+
+- Updated dependencies \[28c2e8b]
+  - @openfn/language-common@2.5.0
+
 ## 1.0.3 - 22 April 2025
 
 ### Patch Changes